android_webview/tools/system_webview_shell/test/data/blink-apis/webmidi/requestmidiaccess.html

<!DOCTYPE html>
<html>
  <script src="../../resources/js-test.js"></script>
  <body>
    <script type="text/javascript">
      description("Test navigator.requestMIDIAccess API in WebView");
      window.jsTestIsAsync = true;

      function onMIDIError(errorMessage) {
        testFailed('requestMIDIAccess error callback, reason: ' + errorMessage);
        finishJSTest();
      }

      function onMIDISuccess(access) {
        window.access = access;
        shouldBeFalse('access.sysexEnabled');
        testPassed('requestMIDIAccess without option succeeded.');
        finishJSTest();
      }

      navigator.requestMIDIAccess().then(
          onMIDISuccess, onMIDIError);
    </script>
  </body>
</html>
