android_webview/tools/system_webview_shell/test/data/blink-apis/webmidi/requestmidiaccess-expected.txt

Test navigator.requestMIDIAccess API in WebView
PASS: access.sysexEnabled is false
PASS: requestMIDIAccess without option succeeded.
TEST FINISHED
